requirements.txt
scipy
bottle
bottle-mysql
mysqlclient
gevent-websocket
gunicorn
psutil
mysql-connector